# Notes-Server/swagger/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "url" : "http://localhost:3000/api/v1"
  } ],
  "info" : {
    "description" : "Add, edit, manager and share notes between users",
    "version" : "0.1-oas3",
    "title" : "Notes"
  },
...
  "paths" : {
...
    "/api/v1/notes/create" : {
      "post" : {
        "tags" : [ "notes" ],
        "summary" : "Create a new note",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "title" : {
                    "type" : "string"
                  },
                  "content" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/notes/delete" : {
      "post" : {
        "tags" : [ "notes" ],
        "summary" : "Delete a note",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find note",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/notes/update" : {
      "put" : {
        "tags" : [ "notes" ],
        "summary" : "Update a note",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find note",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  },
                  "title" : {
                    "type" : "string"
                  },
                  "content" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/notes/fav" : {
      "put" : {
        "tags" : [ "notes" ],
        "summary" : "Toggle favorite status of note",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find note",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/notes/share" : {
      "post" : {
        "tags" : [ "notes" ],
        "summary" : "Share a note with other user",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find note/Cannot find share user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  },
                  "shareuserid" : {
                    "type" : "string"
                  },
                  "sharepermission" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/notes/unshare" : {
      "post" : {
        "tags" : [ "notes" ],
        "summary" : "Unshare a note which is already shared",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find note/Cannot find share user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  },
                  "shareuserid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/notes/" : {
      "post" : {
        "tags" : [ "notes" ],
        "summary" : "Get all notes created by a user",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Note"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "No notes found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/notes/note" : {
      "post" : {
        "tags" : [ "notes" ],
        "summary" : "Get details of a particular note",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Note"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "No notes found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "noteid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Response" : {
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Note" : {
        "properties" : {
          "noteId" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "permission" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "sharedNotes" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "noteId" : {
                  "type" : "string"
                },
                "userId" : {
                  "type" : "string"
                },
                "permission" : {
                  "type" : "string"
                }
              }
            }
          },
          "parentNoteId" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "string"
          },
          "remainderId" : {
            "type" : "string"
          },
          "favorite" : {
            "type" : "boolean"
          },
          "createdOn" : {
            "type" : "string"
          },
          "modifiedOn" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}